+++ io-service/src/main/java/com/example/ioservice/controller/IOController.java
@@ -7,6 +7,7 @@ import com.example.ioservice.dto.WishlistDTOs.WishlistAddDTO;
 import com.example.ioservice.dto.WishlistDTOs.WishlistDTO;
 import com.example.ioservice.dto.WishlistItemDTOs.WishlistItemAddDTO;
 import com.example.ioservice.dto.WishlistItemDTOs.WishlistItemDTO;
+import com.example.ioservice.dto.*;
 import com.example.ioservice.model.ProductModel;
 import com.example.ioservice.repository.ProductRepository;
 import com.example.ioservice.service.IOService;
@@ -108,5 +109,33 @@ public class IOController {
     @PostMapping("/review/delete/{id}")
     public void deleteReview(@PathVariable("id") Long id, @RequestParam("userId") Long userId) {
         ioService.deleteReview(id, userId);
+    @PostMapping("/add_product_to_cart")
+    public String addProductToCart(@RequestBody AddProductToCartDto addProductToCartDto) {
+        return ioService.addProductToCart(addProductToCartDto);
+    }
+
+    @PutMapping("/remove_one_product_from_cart")
+    public String removeOneProduct(@RequestBody RemoveProductFromCartDto removeProductFromCartDto) {
+        return ioService.removeOneProduct(removeProductFromCartDto);
+    }
+
+    @PostMapping("/remove_product_from_cart")
+    public String removeProduct(@RequestBody RemoveProductFromCartDto removeProductFromCartDto) {
+        return ioService.removeProduct(removeProductFromCartDto);
+    }
+
+    @GetMapping("/get_cart_total/{userId}")
+    public Float getCartTotal(@PathVariable String userId) {
+        return ioService.getCartTotal(Long.valueOf(userId));
+    }
+
+    @PostMapping("/make_order")
+    public String makeOrder(@RequestBody OrderDto orderDto) {
+        return ioService.makeOrder(orderDto.userId());
+    }
+
+    @GetMapping("/get_orders_history/{userId}")
+    public List<OrderHistoryDto> getOrdersHistory(@PathVariable String userId) {
+        return ioService.getOrdersHistory(Long.valueOf(userId));
     }
 }