+++ io-service/src/main/java/com/example/ioservice/dto/OrderHistoryDto.java
@@ -1,0 +1,6 @@
+package com.example.ioservice.dto;
+
+import java.util.List;
+
+public record OrderHistoryDto(Long orderId, Float total, String status, List<OrderedProductDto> products) {
+}