+++ io-service/src/main/java/com/example/ioservice/dto/RemoveProductFromCartDto.java
@@ -1,0 +1,4 @@
+package com.example.ioservice.dto;
+
+public record RemoveProductFromCartDto(Long userId, Long productId) {
+}