+++ io-service/src/main/java/com/example/ioservice/model/CartModel.java
@@ -1,0 +1,21 @@
+package com.example.ioservice.model;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Table(name= "cart")
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+@Data
+public class CartModel {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long cartId;
+    private Long userId;
+    private String status;
+}