+++ io-service/src/main/java/com/example/ioservice/repository/CartRepository.java
@@ -1,0 +1,11 @@
+package com.example.ioservice.repository;
+
+import com.example.ioservice.model.CartModel;
+import com.example.ioservice.model.ProductModel;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface CartRepository extends JpaRepository<CartModel, Long> {
+    Optional<CartModel> findByUserIdAndStatus(Long userId, String status);
+}