+++ market-service/src/main/java/com/example/marketservice/dto/OrderedProductDto.java
@@ -1,0 +1,4 @@
+package com.example.marketservice.dto;
+
+public record OrderedProductDto(Long productId, String productName, Float price, Long quantity) {
+}