+++ CHANGELOG.md
@@ -9,6 +9,15 @@
 - Add and delete review to product
 - Get reviews by user or by product
 
+## 0.0.5 - 2025-05-03
+
+### @daria-tanasie
+
+- Add products to cart functionality
+- Added methods for managing the cart (remove, add products)
+- Added order functionality for the cart products
+- Added functionality for order total price
+- Added functionality for orders history view
 
 ## 0.0.4 - 2026-05-01